Add mint CNFT
